sweet alert added in project deletion

+++ client/src/app/project/project.component.ts
@@ -3,7 +3,7 @@ import { Router } from '@angular/router';
 import { Component, OnInit, OnChanges } from '@angular/core';
 import { SendHttpRequestService } from './../send-http-request.service';
 import { EmployeeService } from '../services/employee.service';
-
+import swal from 'sweetalert2';
 @Component({
   selector: "app-project",
   templateUrl: "./project.component.html",
@@ -87,14 +87,48 @@ export class ProjectComponent implements OnInit, OnChanges {
   }
 
   deleteProject(id: any) {
-    console.log(id);
-    this.employeeService.deleteProject(id).subscribe(res => {
-      this.message = res.payload.message;
-      setTimeout(() => {
-        this.message = null;
-      }, 5000);
-      this.projectsArray = this.projectsArray.filter(project =>project.project._id != id);
-    });
+    const swalWithBootstrapButtons = swal.mixin({
+      customClass: {
+        confirmButton: 'btn btn-success',
+        cancelButton: 'btn btn-danger'
+      },
+      buttonsStyling: false
+    })
+    
+    swalWithBootstrapButtons.fire({
+      title: 'Are you sure?',
+      text: "You won't be able to revert this!",
+      icon: 'warning',
+      showCancelButton: true,
+      confirmButtonText: 'Yes, delete it!',
+      cancelButtonText: 'No, cancel!',
+      reverseButtons: true
+    }).then((result) => {
+      if (result.value) {
+        this.employeeService.deleteProject(id).subscribe(res => {
+          this.message = res.payload.message;
+          setTimeout(() => {
+            this.message = null;
+          }, 5000);
+          this.projectsArray = this.projectsArray.filter(project =>project.project._id != id);
+        });
+         swalWithBootstrapButtons.fire(
+          'Deleted!',
+          'Your project data has been deleted.',
+          'success'
+        )
+      } else if (
+        /* Read more about handling dismissals below */
+        result.dismiss === swal.DismissReason.cancel
+      ) {
+        swalWithBootstrapButtons.fire(
+          'Cancelled',
+          'Your projectdata is safe :)',
+          'error'
+        )
+      }
+    }) 
+    
   }
   logout() {
     this._service.deletetoken();
